refactor enums and types to their own files
