+++ src/constants.ts
@@ -1,0 +1,6 @@
+export enum DispatchActions {
+  ADD_ITEM = 'add_item',
+  ADD_PARTY_MEMBER = 'add_party_member',
+  SET_TIP = 'set_tip',
+  SET_TAX = 'set_tax',
+}
